+++ flake.nix
@@ -368,6 +368,20 @@
           titlesec tabulary varwidth framed fancyvrb float wrapfig parskip
           upquote capt-of needspace etoolbox booktabs;
       };
+
+      artiq-frontend-dev-wrappers = pkgs.runCommandNoCC "artiq-frontend-dev-wrappers" {}
+        ''
+        mkdir -p $out/bin
+        for program in ${self}/artiq/frontend/*.py; do
+          if [ -x $program ]; then
+            progname=`basename -s .py $program`
+            outname=$out/bin/$progname
+            echo "#!${pkgs.bash}/bin/bash" >> $outname
+            echo "exec python3 -m artiq.frontend.$progname \"\$@\"" >> $outname
+            chmod 755 $outname
+          fi
+        done
+        '';
     in rec {
       packages.x86_64-linux = {
         inherit pythonparser qasync artiq;
@@ -431,8 +445,9 @@
       defaultPackage.x86_64-linux = pkgs.python3.withPackages(ps: [ packages.x86_64-linux.artiq ]);
 
       # Main development shell with everything you need to develop ARTIQ on Linux.
-      # ARTIQ itself is not included in the environment, you can make Python use the current sources using e.g.
-      # export PYTHONPATH=`pwd`:$PYTHONPATH
+      # The current copy of the ARTIQ sources is added to PYTHONPATH so changes can be tested instantly.
+      # Additionally, executable wrappers that import the current ARTIQ sources for the ARTIQ frontends
+      # are added to PATH.
       devShells.x86_64-linux.default = pkgs.mkShell {
         name = "artiq-dev-shell";
         buildInputs = [
@@ -442,6 +457,8 @@
           pkgs.llvmPackages_14.clang-unwrapped
           pkgs.llvm_14
           pkgs.lld_14
+          pkgs.git
+          artiq-frontend-dev-wrappers
           # To manually run compiler tests:
           pkgs.lit
           outputcheck
@@ -457,6 +474,7 @@
           export LIBARTIQ_SUPPORT=`libartiq-support`
           export QT_PLUGIN_PATH=${pkgs.qt5.qtbase}/${pkgs.qt5.qtbase.dev.qtPluginPrefix}:${pkgs.qt5.qtsvg.bin}/${pkgs.qt5.qtbase.dev.qtPluginPrefix}
           export QML2_IMPORT_PATH=${pkgs.qt5.qtbase}/${pkgs.qt5.qtbase.dev.qtQmlPrefix}
+          export PYTHONPATH=`git rev-parse --show-toplevel`:$PYTHONPATH
         '';
       };